requirements.txt
# VLA Robot Learning Framework Requirements
# Core deep learning framework
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Computer vision
opencv-python>=4.5.0
Pillow>=8.3.0
imageio>=2.15.0

# Data processing
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress bars
tqdm>=4.62.0

# Configuration management
pyyaml>=5.4.0
hydra-core>=1.1.0

# Logging
loguru>=0.5.0

# Type checking
typing-extensions>=3.10.0

# Path handling
pathlib2>=2.3.0

# JSON processing
orjson>=3.6.0

# Physics simulation
mujoco>=3.1.0
mujoco-mjx>=0.2.0
dm-control>=1.0.0
gymnasium>=0.28.0

# ROS integration (optional)
# rospy>=1.15.0
# rospkg>=1.4.0

# Web and API
flask>=2.0.0
requests>=2.25.0

# Database
sqlalchemy>=1.4.0
sqlite3

# Version control
gitpython>=3.1.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code quality
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Development tools
jupyter>=1.0.0
ipython>=7.0.0

# Performance optimization
numba>=0.56.0

# GPU acceleration (optional)
# cupy-cuda11x>=10.0.0

# Model deployment
onnx>=1.10.0
onnxruntime>=1.8.0

# Experiment tracking
tensorboard>=2.8.0
wandb>=0.12.0

# Hyperparameter optimization
optuna>=2.10.0

# Data version control
dvc>=2.10.0

# Containerization
docker>=5.0.0

# Build tools
setuptools>=60.0.0
wheel>=0.37.0

# Security
cryptography>=3.4.0